+++ packages/expo-modules-core/ios/Swift/ExpoBridgeModule.m
@@ -1,0 +1,5 @@
+#import <React/RCTBridgeModule.h>
+
+@interface RCT_EXTERN_MODULE(ExpoBridgeModule, NSObject)
+
+@end