+++ packages/expo-modules-core/ios/Tests/ExpoModulesSpec.swift
@@ -7,7 +7,6 @@ import ExpoModulesTestCore
 class ExpoModulesSpec: ExpoSpec {
   override func spec() {
     let appContext = AppContext.create()
-    let interopBridge = SwiftInteropBridge(appContext: appContext)
     let runtime = appContext.runtime
     let testModuleName = "TestModule"
     let testFunctionName = "testFunction"
@@ -17,8 +16,6 @@ class ExpoModulesSpec: ExpoSpec {
     ]
 
     beforeSuite {
-      try! appContext.installExpoModulesHostObject(interopBridge)
-
       appContext.moduleRegistry.register(holder: mockModuleHolder(appContext) {
         Name(testModuleName)